requirements.txt
Flask==2.0.1
gunicorn==20.1.0
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
xgboost==1.4.2
psycopg2-binary==2.9.1